+++ src/main/scala/org/podval/xml/html/Givens.scala
@@ -1,3 +1,0 @@
-package org.podval.xml.html
-
-given org.podval.xml.Html.type = org.podval.xml.Html