+++ src/main/scala/org/podval/xml/scalaxml/Givens.scala
@@ -1,3 +1,0 @@
-package org.podval.xml.scalaxml
-
-given org.podval.xml.ScalaXml.type = org.podval.xml.ScalaXml