+++ src/test/scala/org/podval/xml/ScalaXmlSpec.scala
@@ -1,6 +1,6 @@
 package org.podval.xml
 
-import org.podval.xml.scalaxml.given
+import ScalaXml.given
 import org.scalatest.funsuite.AnyFunSuite
 
 final class ScalaXmlSpec extends AnyFunSuite: