+++ src/test/scala/org/podval/xml/XmlNamespaceSpec.scala
@@ -1,6 +1,6 @@
 package org.podval.xml
 
-import org.podval.xml.scalaxml.given
+import ScalaXml.given
 import org.scalatest.funsuite.AnyFunSuite
 import zio.blocks.schema.xml.XmlName